refactor: rename adoptium-java->java-version
