+++ lib/modules/datasource/java-version/__fixtures__/jre.json
