+++ lib/modules/datasource/java-version/__fixtures__/page.json
